+++ web/src/pages/Event.tsx
@@ -1,10 +1,10 @@
-import Navbar from "../Components/Navbar";
+import Navbar from "../component/Navbar";
 
 export default function Event() {
-  return (
-    <div>
-      <Navbar />
-      <h1> Events Page</h1>
-    </div>
-  )
+	return (
+		<div>
+			<Navbar />
+			<h1> Events Page</h1>
+		</div>
+	);
 }
