+++ web/src/pages/Home.tsx
@@ -1,14 +1,12 @@
-import Navbar from "../Components/Navbar";
+import Navbar from "../component/Navbar";
 
 export default function Home() {
-  return (
-    <>
-    <Navbar />
-    <div className="container px-4 pt-24 mx-auto content-center">
-      <h1 className=" text-center text-3xl font-bold text-gray-900 sm:text-4xl dark:text-white">
-        Home Page
-      </h1>
-    </div>
-    </>
-  );
+	return (
+		<>
+			<Navbar />
+			<div className="container px-4 pt-24 mx-auto content-center">
+				<h1 className=" text-center text-3xl font-bold text-gray-900 sm:text-4xl dark:text-white">Home Page</h1>
+			</div>
+		</>
+	);
 }
