+++ web/src/pages/SignUp.tsx
@@ -1,11 +1,10 @@
-import Navbar from "../Components/Navbar";
+import Navbar from "../component/Navbar";
 
 export default function SignUp() {
-    return (
-      <div>
-        <Navbar />
-        <h1> Sign Up Page</h1>
-      </div>
-    )
+	return (
+		<div>
+			<Navbar />
+			<h1> Sign Up Page</h1>
+		</div>
+	);
 }
-
